# docs/openapi.yaml
openapi: 3.0.3
info:
  title: 兜行 Douxing API
  description: |
    兜行平台 REST API（`packages/server`）。

    - 统一响应：`{ code, message, messageKey?, data }`，`code === 0` 为成功
    - 鉴权：Header `Authorization: Bearer <Access Token JWT>`；过期后客户端用 Refresh Token 调 `POST /api/auth/refresh`（见 [双Token认证与无感刷新.md](./双Token认证与无感刷新.md)）
    - 国际化：Header `Accept-Language: zh-CN | en-US`
    - 人类可读说明见 [API接口文档.md](./API接口文档.md)
  version: 1.4.1
  contact:
    name: 兜行项目
servers:
  - url: http://localhost:3000
    description: 本地开发
  - url: https://api.yxtao.site
    description: 生产环境
...
security:
  - bearerAuth: []
paths:
...
  /api/users/me:
    get:
      tags: [用户]
      summary: 获取当前用户资料
      operationId: usersMeGet
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserInfo'
    put:
      tags: [用户]
      summary: 更新用户资料
      operationId: usersMePut
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserProfileRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserInfo'
...
  /api/users/me/avatar:
    post:
      tags: [用户]
      summary: 上传头像
      operationId: usersMeAvatar
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: 图片，≤ 2MB
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserInfo'
...
  /api/routes/{id}:
    get:
      tags: [路线]
      summary: 路线详情
      operationId: routesDetail
      parameters:
        - $ref: '#/components/parameters/RouteId'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    put:
      tags: [路线]
      summary: 编辑草稿路线
      operationId: routesUpdateDraft
      parameters:
        - $ref: '#/components/parameters/RouteId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRouteDraftRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
  /api/share/routes/{id}:
    get:
      tags: [分享]
      summary: 公开路线只读摘要
      security: []
      operationId: shareRouteDetail
      parameters:
        - $ref: '#/components/parameters/RouteId'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "登录后获得的 Access Token（JWT），填入 Authorization: Bearer <token>"
  parameters:
...
    RouteId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
...
  schemas:
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: 成功
        messageKey:
          type: string
          example: api.ok
        data:
          nullable: true
...
    UserInfo:
      type: object
      properties:
        id: { type: integer }
        username: { type: string }
        nickname: { type: string }
        avatar: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        email: { type: string, nullable: true }
        interestTags:
          type: array
          items: { type: string }
          nullable: true
        memberLevel: { type: integer }
        status: { type: integer }
        roles:
          type: array
          items: { type: string }
...
    UpdateUserProfileRequest:
      type: object
      properties:
        nickname: { type: string, minLength: 1, maxLength: 64 }
        avatar: { type: string, nullable: true }
        email: { type: string, nullable: true }
        interestTags:
          type: array
          items: { type: string }
...
    UpdateRouteDraftRequest:
      type: object
      properties:
        name: { type: string }
        description: { type: string, nullable: true }
        budgetRange: { type: string, nullable: true }
        days: { type: integer }
        interestTags:
          type: array
          items: { type: string }
        routeDetail:
          type: object
...